[cleanup] Moved python unittest to its folder under name factory
