+++ unittest/bindings/test_actions.py
@@ -7,7 +7,7 @@ import numpy as np
 import pinocchio
 
 import crocoddyl
-from crocoddyl.utils import (
+from factory import (
     DifferentialFreeFwdDynamicsModelDerived,
     DifferentialLQRModelDerived,
     IntegratedActionModelEulerDerived,
